+++ resources/metadata.rb
@@ -1,7 +1,7 @@
-name             'example'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
-license          'All rights reserved'
-description      'Installs/Configures cookbook-example'
+name             'f2k'
+maintainer       'Alberto Rodriguez'
+maintainer_email 'arodriguez@redborder'
+license          'AGPL v3'
+description      'Installs/Configures cookbook-f2k'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.0.1'