+++ resources/resources/config.rb
@@ -1,4 +1,4 @@
-# Cookbook Name:: example
+# Cookbook Name:: f2k
 #
 # Resource:: config
 #
@@ -6,8 +6,10 @@
 actions :add, :remove, :register, :deregister
 default_action :add
 
-attribute :mystring, :kind_of => String, :default => "string example"
-attribute :myinteger, :kind_of => Fixnum, :default => 1
-attribute :myarray, :kind_of => Array, :default => ["val1"]
-attribute :myhash, :kind_of => Object, :default => {"val1" => "1"}
-attribute :myboolean, :kind_of => [TrueClass, FalseClass], :default => true
+attribute :cores, :kind_of => Integer, :default => 1
+attribute :memory_kb, :kind_of => Integer, :default => 102400
+attribute :enrichment_enabled, :kind_of => [TrueClass, FalseClass], :default => true
+attribute :cache_dir, :kind_of => String, :default => "/var/cache/f2k"
+attribute :config_dir, :kind_of => String, :default => "/etc/f2k"
+attribute :user, :kind_of => String, :default => "f2k"
+attribute :sensors, :kind_of => Hash, :default => []